requirements.txt
Django==1.5.2
Fabric==1.7.0
MySQL-python==1.2.5
Pillow
South==0.8.2
Unipath==1.0
argparse==1.2.1
boto==2.22.1
django-bootstrap-form==2.0.6
django-debug-toolbar==0.9.4
django-extensions==1.1.1
django-flashpolicies==1.5
django-model-utils==1.4.0
django-ses==0.4.1
flup==1.0.2
paramiko==1.11.0
pycrypto==2.6
six==1.3.0
wsgiref==0.1.2
